Compute proper set of supported ciphers for EncryptorTest unit test
The "openssl list-cipher-commands" command will report some ciphers, such as
"base64" or "zlib", which are supported by the native OpenSSL commands but which
are not supported by the Ruby OpenSSL class. What Ruby supports depends on how
it was installed/compiled. The safest way to determine the set of ciphers
supported on the current machine is to calculate the intersection of this set
with the contents of the OpenSSL::Cipher.ciphers array.