rescue errors coming out of OpenSSL::Cipher and reraise our own
There are a couple motivations for this change:

First as library authors we should help our clients out by containing
any exceptions raised by our underlying libraries and either handling
them or making them easy to rescue by raising something that comes from
us.

Second, there is only one exception that `OpenSSL::Cipher` is
currently written to raise and that is: `CipherError`. Unfortunately
this makes it really difficult to handle these exceptions so we can do 
that in this central location and reraise something more descriptive.

Signed-off-by: Tyler Pickett <t.pickett66@gmail.com>