Introduce `Encryptor::Cipher`
This class is a wrapper around an instance of `OpenSSL::Cipher::Cipher`
with a configured set of options e.g. `:algorithm`, `:key`, `:salt`,
etc. This object responds to both `encrypt` and `decrypt` which accept a
`value` argument.

This allows `Encryptor` itself to act more as an interface which can
be mixed into other classes to add encryption behavior since it's a
module. It also removes the protected method `crypt` from the module and
pushes that logic down into the `Encryptor::Cipher` class. This cleans
up the `Encryptor` api, defining only the `encrypt` and `decrypt`
methods when included in a class.

This also allows us to `include Encryptor` within `Encrytor::String`
instead of referencing the top level `Encryptor` module inside of the
string's `encrypt` and `decrypt` methods. We couldn't do this previously
since `Encryptor#crypt` was defined and conflicts with Ruby's standard
`String#crypt` method. This simplifies the `Encryptor::String` module
quite a bit.

module Encryptor::String
  include Encryptor

  def encrypt(options = {}, &block)
    super(self, options, &block)
  end

  def decrypt(options = {}, &block)
    super(self, options, &block)
  end
end